+++ lib/src/services/api/autotest_api_service.dart
@@ -8,12 +8,11 @@ import 'package:adapters_flutter/src/managers/log_manager.dart';
 import 'package:adapters_flutter/src/models/api/autotest_api_model.dart';
 import 'package:adapters_flutter/src/models/api/workitem_api_model.dart';
 import 'package:adapters_flutter/src/models/config_model.dart';
+import 'package:adapters_flutter/src/models/exception_model.dart';
 import 'package:adapters_flutter/src/models/test_result_model.dart';
-import 'package:adapters_flutter/src/services/validation_service.dart';
-import 'package:dio/dio.dart';
+import 'package:http/http.dart';
 import 'package:meta/meta.dart';
 
-final _dio = Dio();
 final _logger = getLogger();
 
 @internal
@@ -28,25 +27,30 @@ Future<AutoTestFullModel?> createAutoTestAsync(
       'Authorization': 'PrivateToken ${config.privateToken}'
     };
 
-    final options = Options(headers: headers);
-    final url = Uri.parse('${config.url}/api/v2/autoTests');
+    final request = Request(
+        'POST', Uri.tryParse('${config.url}/api/v2/autoTests') ?? Uri());
 
-    final body = toCreateAutoTestRequestModel(config.projectId, testResult);
-    body.shouldCreateWorkItem = config.automaticCreationTestCases ?? false;
-    final data = json.encode(body);
+    final requestBody =
+        toCreateAutoTestRequestModel(config.projectId, testResult);
+    requestBody.shouldCreateWorkItem =
+        config.automaticCreationTestCases ?? false;
+    request.body = json.encode(requestBody);
+    request.headers.addAll(headers);
 
-    final response = await _dio.postUri(url, data: data, options: options);
-    final exception = getResponseValidationException(response);
+    final streamedResponse = await request.send();
+    final response = await Response.fromStream(streamedResponse);
 
-    if (exception != null) {
+    if (response.statusCode < 200 || response.statusCode > 299) {
+      final exception = TmsApiException(
+          'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
       _logger.i('$exception.');
 
       return autoTest;
     }
 
-    autoTest =
-        AutoTestFullModel.fromJson(response.data as Map<String, dynamic>);
-  } on DioException catch (exception, stacktrace) {
+    final body = jsonDecode(response.body) as Map<String, dynamic>;
+    autoTest = AutoTestFullModel.fromJson(body);
+  } catch (exception, stacktrace) {
     _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
   }
 
@@ -65,10 +69,13 @@ Future<AutoTestFullModel?> getAutoTestByExternalIdAsync(
       'Authorization': 'PrivateToken ${config.privateToken}'
     };
 
-    final options = Options(headers: headers);
-    final url = Uri.parse(
-        '${config.url}/api/v2/autoTests/search?SearchField=externalId&SearchValue=$externalId');
-    final data = json.encode({
+    final request = Request(
+        'POST',
+        Uri.tryParse(
+                '${config.url}/api/v2/autoTests/search?SearchField=externalId&SearchValue=$externalId') ??
+            Uri());
+
+    request.body = json.encode({
       'filter': {
         'projectIds': ['${config.projectId}'],
         'isDeleted': false,
@@ -80,19 +87,23 @@ Future<AutoTestFullModel?> getAutoTestByExternalIdAsync(
       }
     });
 
-    final response = await _dio.postUri(url, data: data, options: options);
-    final exception = getResponseValidationException(response);
+    request.headers.addAll(headers);
 
-    if (exception != null) {
+    final streamedResponse = await request.send();
+    final response = await Response.fromStream(streamedResponse);
+
+    if (response.statusCode < 200 || response.statusCode > 299) {
+      final exception = TmsApiException(
+          'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
       _logger.i('$exception.');
 
       return autoTest;
     }
 
-    autoTest = (response.data as List<dynamic>)
-        .map((autoTest) => AutoTestFullModel.fromJson(autoTest))
-        .singleOrNull;
-  } on DioException catch (exception, stacktrace) {
+    final body =
+        (jsonDecode(response.body) as List).cast<Map<String, dynamic>>();
+    autoTest = AutoTestFullModel.fromJson(body.single);
+  } catch (exception, stacktrace) {
     _logger.d('$exception${Platform.lineTerminator}$stacktrace.');
   }
 
@@ -110,22 +121,25 @@ Future<List<String>> getWorkItemsGlobalIdsLinkedToAutoTestAsync(
       'Authorization': 'PrivateToken ${config.privateToken}'
     };
 
-    final options = Options(headers: headers);
-    final url = Uri.parse(
-        '${config.url}/api/v2/autoTests/$autoTestId/workItems?isDeleted=false');
+    final request = Request(
+        'GET',
+        Uri.parse(
+            '${config.url}/api/v2/autoTests/$autoTestId/workItems?isDeleted=false'));
+    request.headers.addAll(headers);
 
-    final response = await _dio.getUri(url, options: options);
-    final exception = getResponseValidationException(response);
+    final streamedResponse = await request.send();
+    final response = await Response.fromStream(streamedResponse);
 
-    if (exception != null) {
+    if (response.statusCode < 200 || response.statusCode > 299) {
+      final exception = TmsApiException(
+          'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
       _logger.i('$exception.');
-
-      return globalIds;
     }
 
-    globalIds.addAll((response.data as List<dynamic>)
-        .map((final workItem) => (workItem['globalId'] as int).toString()));
-  } on DioException catch (exception, stacktrace) {
+    final body = jsonDecode(response.body) as List<dynamic>;
+    globalIds.addAll(
+        body.map((final workItem) => (workItem['globalId'] as int).toString()));
+  } catch (exception, stacktrace) {
     _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
   }
 
@@ -143,18 +157,22 @@ Future<void> linkWorkItemsToAutoTestAsync(final String? autoTestId,
         'Authorization': 'PrivateToken ${config.privateToken}'
       };
 
-      final options = Options(headers: headers);
-      final url =
-          Uri.parse('${config.url}/api/v2/autoTests/$autoTestId/workItems');
-      final data = json.encode(WorkItemLinkRequestModel(id));
+      final request = Request(
+          'POST',
+          Uri.tryParse(
+                  '${config.url}/api/v2/autoTests/$autoTestId/workItems') ??
+              Uri());
+      request.body = json.encode(WorkItemLinkRequestModel(id));
+      request.headers.addAll(headers);
 
-      final response = await _dio.postUri(url, data: data, options: options);
-      final exception = getResponseValidationException(response);
+      final response = await request.send();
 
-      if (exception != null) {
+      if (response.statusCode < 200 || response.statusCode > 299) {
+        final exception = TmsApiException(
+            'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
         _logger.i('$exception.');
       }
-    } on DioException catch (exception, stacktrace) {
+    } catch (exception, stacktrace) {
       _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
     }
   }
@@ -171,17 +189,20 @@ Future<void> unlinkAutoTestFromWorkItemsAsync(final String? autoTestId,
         'Authorization': 'PrivateToken ${config.privateToken}'
       };
 
-      final options = Options(headers: headers);
-      final url = Uri.parse(
-          '${config.url}/api/v2/autoTests/$autoTestId/workItems?workItemId=$id');
+      final request = Request(
+          'DELETE',
+          Uri.parse(
+              '${config.url}/api/v2/autoTests/$autoTestId/workItems?workItemId=$id'));
+      request.headers.addAll(headers);
 
-      final response = await _dio.deleteUri(url, options: options);
-      final exception = getResponseValidationException(response);
+      final response = await request.send();
 
-      if (exception != null) {
+      if (response.statusCode < 200 || response.statusCode > 299) {
+        final exception = TmsApiException(
+            'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
         _logger.i('$exception.');
       }
-    } on DioException catch (exception, stacktrace) {
+    } catch (exception, stacktrace) {
       _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
     }
   }
@@ -197,18 +218,21 @@ Future<void> updateAutoTestAsync(
       'Authorization': 'PrivateToken ${config.privateToken}'
     };
 
-    final options = Options(headers: headers);
-    final url = Uri.parse('${config.url}/api/v2/autoTests');
-    final data =
-        json.encode(toUpdateAutoTestRequestModel(config.projectId, testResult));
+    final request =
+        Request('PUT', Uri.tryParse('${config.url}/api/v2/autoTests') ?? Uri());
+    final requestBody =
+        toUpdateAutoTestRequestModel(config.projectId, testResult);
+    request.body = json.encode(requestBody);
+    request.headers.addAll(headers);
 
-    final response = await _dio.putUri(url, data: data, options: options);
-    final exception = getResponseValidationException(response);
+    final response = await request.send();
 
-    if (exception != null) {
+    if (response.statusCode < 200 || response.statusCode > 299) {
+      final exception = TmsApiException(
+          'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
       _logger.i('$exception.');
     }
-  } on DioException catch (exception, stacktrace) {
+  } catch (exception, stacktrace) {
     _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
   }
 }