+++ CHANGELOG.md
@@ -3,11 +3,20 @@
 All notable changes to this project will be documented in this file.  
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
-## [1.6.1] - 2024-07-08
+## [1.7.0] - 2024-07-08
 
-### Changed
+### Added
 
-- 'flutter_test' included in 'adapters_flutter' package.
+- readme requirements.
+- test action.
+- test examples.
+
+### Fixed
+
+- addLink & addLinks functions.
+- check action.
+- externalId resolver.
+- test name.
 
 ## [1.6.0] - 2024-07-03